Fix removing script references when script file doesn't exist (in case of bad database import, without associated scripts)

+++ sources/server/automation/interpreter/Manager.cpp
@@ -237,6 +237,8 @@ namespace automation
       {
          if (doBackup)
          {
+            if (!boost::filesystem::exists(scriptPath))
+               return; // To remove reference to unexisting scripts (in case of database import, without associated scripts)
             const auto backupPath(scriptPath + ".bak");
             boost::filesystem::remove_all(backupPath);
             boost::filesystem::rename(scriptPath, backupPath);
